+++ test/bract/ring/inducer_test.clj
@@ -30,30 +30,25 @@
   handler)
 
 
-(deftest test-ctx-apply-wrappers
-  (testing "happy cases"
-    (let [context {:bract.ring/ring-handler identity
-                   :bract.ring/wrappers     [wrapper-inc wrapper-add2]}]
+(deftest test-apply-wrappers
+  (testing "happy cases, wrapper fns"
+    (let [context {:bract.ring/ring-handler identity}]
       (vreset! holder 0)
-      (let [new-context (inducer/ctx-apply-wrappers context)
+      (let [new-context (inducer/apply-wrappers context [wrapper-inc wrapper-add2])
             new-handler (ring-config/ctx-ring-handler new-context)]
-        (is (contains? new-context :bract.ring/wrappers))
+        (is (contains? new-context :bract.ring/ring-handler))
         (is (= :foo (new-handler :foo))))
       (is (= 3 @holder))))
-  (testing "missing/bad keys"
-    (is (thrown? IllegalArgumentException (inducer/ctx-apply-wrappers {})))))
-
-
-(deftest test-cfg-apply-wrappers
-  (testing "happy cases"
-    (let [context {:bract.ring/ring-handler identity
-                   :bract.core/config {"bract.ring.wrappers" ['bract.ring.inducer-test/wrapper-inc
-                                                              'bract.ring.inducer-test/wrapper-add2]}}]
+  (testing "happy cases, wrapper names"
+    (let [context {:bract.ring/ring-handler identity}]
       (vreset! holder 0)
-      (let [new-context (inducer/cfg-apply-wrappers context)
+      (let [new-context (inducer/apply-wrappers context '[bract.ring.inducer-test/wrapper-inc
+                                                          bract.ring.inducer-test/wrapper-add2])
             new-handler (ring-config/ctx-ring-handler new-context)]
-        (is (contains? (core-config/ctx-config new-context) "bract.ring.wrappers"))
+        (is (contains? new-context :bract.ring/ring-handler))
         (is (= :foo (new-handler :foo))))
       (is (= 3 @holder))))
-  (testing "missing/bad keys"
-    (is (thrown? IllegalArgumentException (inducer/cfg-apply-wrappers {})))))
+  (testing "empty wrappers collection"
+    (inducer/apply-wrappers {} []))
+  (testing "missing Ring handler"
+    (is (thrown? IllegalArgumentException (inducer/apply-wrappers {} [(fn [x y] x)])))))