+++ CHANGELOG.md
@@ -1,7 +1,18 @@
-# TODO and Change Log
+# bract.ring TODO and Change Log
 
 ## TODO
 None
+
+## 0.3.0 / 2017-June-??
+### Changed
+- Use bract.core 0.3
+- Add inducer `bract.ring.inducer/apply-wrappers` accepting Ring wrappers
+- [BREAKING CHANGE] Remove unparameterized wrapper-applying inducers in favor of `apply-wrappers`
+  - Remove inducer `bract.ring.inducer/ctx-apply-wrappers`
+  - Remove inducer `bract.ring.inducer/cfg-apply-wrappers`
+  - Remove config definition `bract.ring.config/ctx-wrappers`
+  - Remove config definition `bract.ring.config/cfg-wrappers`
+
 
 ## 0.2.0 / 2017-June-04
 ### Changed